Bicaws 1097 preprod test target

+++ helpers/MockPNCHelper.js
@@ -1,6 +1,7 @@
 const axios = require("axios").default;
 const fs = require("fs").promises;
 const Poller = require("../utils/Poller");
+const realPNC = process.env.REAL_PNC && process.env.REAL_PNC === "true";
 
 class MockPNCHelper {
   constructor(options) {
@@ -8,6 +9,7 @@ class MockPNCHelper {
   }
 
   async addMock(matchRegex, response, count = null) {
+    if (realPNC) return;
     const data = { matchRegex, response, count };
     const resp = await axios.post(`http://${this.options.host}:${this.options.port}/mocks`, data);
     if (resp.status < 200 || resp.status >= 300) {
@@ -17,6 +19,7 @@ class MockPNCHelper {
   }
 
   async getMock(id) {
+    if (realPNC) return;
     const resp = await axios.get(`http://${this.options.host}:${this.options.port}/mocks/${id}`, {
       validateStatus: () => true
     });
@@ -27,6 +30,7 @@ class MockPNCHelper {
   }
 
   async getRequests() {
+    if (realPNC) return;
     const resp = await axios.get(`http://${this.options.host}:${this.options.port}/requests`, {
       validateStatus: () => true
     });
@@ -37,6 +41,7 @@ class MockPNCHelper {
   }
 
   async getMocks() {
+    if (realPNC) return;
     const resp = await axios.get(`http://${this.options.host}:${this.options.port}/mocks`, {
       validateStatus: () => true
     });
@@ -47,6 +52,7 @@ class MockPNCHelper {
   }
 
   async awaitMockRequest(id, timeout = 40000) {
+    if (realPNC) return;
     const action = () => this.getMock(id);
 
     const condition = (mock) => mock && mock.requests && mock.requests.length > 0;
@@ -64,6 +70,7 @@ class MockPNCHelper {
   }
 
   async clearMocks() {
+    if (realPNC) return;
     const response = await axios.delete(`http://${this.options.host}:${this.options.port}/mocks`);
     if (response.status !== 204) {
       throw new Error("Error clearing mocks in PNC Emulator");
@@ -71,11 +78,13 @@ class MockPNCHelper {
   }
 
   async recordRequests() {
+    if (realPNC) return;
     const requests = await this.getRequests();
     await fs.writeFile(`${this.options.world.outputDir}/requests.json`, JSON.stringify(requests));
   }
 
   async recordMocks() {
+    if (realPNC) return;
     const mocks = await this.getMocks();
     await fs.writeFile(`${this.options.world.outputDir}/mocks.json`, JSON.stringify(mocks));
   }
